requirements.txt
alembic==0.9.5
backcall==0.1.0
bcrypt==3.1.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
click==6.7
coverage==4.4.1
decorator==4.3.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.0
Flask-MQTT==1.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
ForgeryPy==0.1
gunicorn==19.7.1
idna==2.6
ipykernel==4.8.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
nose==1.3.7
paho-mqtt==1.3.1
parso==0.2.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.2
python-dateutil==2.6.1
python-editor==1.0.3
pyzmq==17.0.0
rauth==0.7.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.13
typing==3.6.4
urllib3==1.22
Werkzeug==0.12.2
